Broke out the defineGoals function from main. #28
